Fix icon issue and identify request origin
Originally, the server responded with a content type of text/plain.
And it could lead to issues when accessed through a web browser.
The state occurs particularly when handling requests for favorite icon.
So, update response type to ensure the web page is well-performed.
Another update is to extract the User-Agent field from requests.
And store it in the http_request_t structure.
Use this information to identify where the request from.
Finally, modify the response.

Change-Id: Ib13b7e08204e804e76bbc2699787fe50c46ca754